+++ src/algorithms/mod.rs
@@ -7,9 +7,12 @@ pub mod line_search;
 
 /// Module containing the BFGS method.
 pub mod bfgs;
+pub use bfgs::BFGS;
 
 /// Module containing the L-BFGS method.
 pub mod lbfgs;
+pub use lbfgs::LBFGS;
 
 /// Module containing the L-BFGS-B method.
 pub mod lbfgsb;
+pub use lbfgsb::LBFGSB;